# Cache the expensive only_us_ascii? check