Is this safe to do?

In my benchmarks I'm seeing from 15-20% speed improvement when building emails with a 40KB html part and 30-40% speed improvement when building emails with large attachments.