@dalibor 2017-06-01
@jeremy Great using `String#ascii_only?`! The only difference is that the original regex excludes `"\x00"` while `ascii_only?` includes it, does it matter? With this change the test suite time goes from 2.3 seconds to 1.6 seconds. And also, I'm seeing significant performance improvements in benchmarks locally building real emails.

```
30 attachments of ~ 10MB each:   17.801965 sec -> 5.149144 sec (~ 3x faster)
10000 html parts of ~ 40KB each: 16.148903 sec -> 9.189533 sec (~ 1x faster)
```