@jeremy 2017-06-01
IMO that was a bug, since we don't want NUL bytes in our US-ASCII text either :)

Very nice improvements!