@dalibor 2017-06-01
It's the other way around, the new `ascii_only?` check will allow NUL bytes in emails I'm afraid. :) If we want to prevent that, we could do `raw_source.ascii_only? && raw_source !~ /\x00/` :/, but that would be slower than `ascii_only?` and faster than the original regex.