@jeremy 2017-06-11
Hmm yes. This seems like an implementation wart, really. NUL is legit for US-ASCII. It's not legal in emails, but that's not what we're validating here, so we can get away with the `ascii_only?`check.